+++ R/zzz.R
@@ -1,11 +1,20 @@
-.osf_api_version <- 2.8
-.wb_api_version  <- 1
-
 .onLoad <- function(libname, pkgname) {
 
   # record personal access token
   env.pat <- Sys.getenv("OSF_PAT")
   if (nzchar(env.pat)) options(osfr.pat = env.pat)
+
+  # setup logger
+  env.log <- Sys.getenv("OSF_LOG")
+  if (nzchar(env.log)) {
+    if (requireNamespace("logger", quietly = TRUE)) {
+      options(osfr.log = env.log)
+      logger::log_appender(logger::appender_file(env.log), namespace = "osfr")
+      logger::log_formatter(logger::formatter_sprintf, namespace = "osfr")
+    } else {
+      warn("The logger package must installed to enable logging")
+    }
+  }
 
   # register dplyr methods
   if (requireNamespace("dplyr", quietly = TRUE)) {
@@ -23,6 +32,12 @@
 .onAttach <- function(libname, pkgname) {
   if (!is.null(getOption("osfr.pat"))) {
     packageStartupMessage("Automatically registered OSF personal access token")
+  }
+
+  if (!is.null(getOption("osfr.log"))) {
+      packageStartupMessage(
+        sprintf("<Logging enabled: %s>", getOption("osfr.log"))
+      )
   }
 
   server <- Sys.getenv("OSF_SERVER")