+++ man-roxygen/progress.R
@@ -1,0 +1,3 @@
+#' @param progress Logical, if `TRUE` progress bars are displayed for each file
+#'   transfer. Mainly useful for transferring large files. For tracking lots of
+#'   small files, setting `verbose = TRUE` is more informative.